refactor!: Refactor exports to reduce indirection
BREAKING CHANGE: `clearAll()` returns undefined instead of the Nock global

There are a lot of places in the Nock codebase where it’s not clear how control passes from one module to another. This tries to improve the situation by using named imports internally in place of some places where default + named imports were combined, and removing a few cases of indirection entirely. Overall the goal is to give the imports a more conventional feeling.

The nock global is now assembled in one place, in `index.js`, instead of being partially assembled in `scope.js` and finished in `index.js`.